Icons: Deprecate `warning` and rename to `cautionFilled`
